% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{combine_variances}
\alias{combine_variances}
\title{Create a dataset of DEPM parameter variances to be passed to Estimate_biomass()}
\usage{
combine_variances(P0, A, R, S, W_F = NULL)
}
\arguments{
\item{P0}{The results of Estimate_P0() with a single value of Z. Multiple values are not yet supported.}

\item{A}{A data.frame of spawning area in meters^2. Can be specified for each Time/Region combination if necessary}

\item{R}{A data.frame of sex ratios. Can be specified for each Time/Region combination if necessary}

\item{S}{A data.frame of spawning fractions. Can be specified for each Time/Region combination if necessary}

\item{W_F}{The results of  Estimate_mean_W_F(). This is optional and only required if using the standard DEPM methods.
Can be specified for each Time/Region combination if necessary}
}
\value{
A data.frame of DEPM parameters compatible with the Estimate_biomass() function
}
\description{
Combine all of the variances determined from DEPM parameter functions in this package. The structure of this object
   is compatible with the Estimate_biomass() function to facilitate its use. If multiple time steps/Regions are included then
   parameter variances specific to those surveys will be combined. Parameters that are used across years (for example a constant sex ratio)
   will automatically be included for each survey even if other parameters are being specified for specific steps/Regions.
}
