% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{combine_wt_class_estimates}
\alias{combine_wt_class_estimates}
\title{Create a dataset of DEPM parameter weight class parameter estimates and variances to be passed to Estimate_biomass()}
\usage{
combine_wt_class_estimates(prop.fem.data, fecundity.data)
}
\arguments{
\item{prop.fem.data}{A dataframe produced from the the Estimate_proportion_female() function}

\item{fecundity.data}{A dataframe produced from the the Estimate_Batch_Fecundity()  function}
}
\value{
A dataframe with the columns Wt_bin, n, Prop, Prop_var, Wt, Fecundity and Fec_var.
    Time and Region are included if included in either dataset
}
\description{
Combine the estimates and variances determined from the Estimate_proportion_female() and Estimate_Batch_Fecundity()
    functions in this package. The structure of this object is compatible with the Estimate_biomass() function to facilitate
    its use. If multiple time steps/Regions are included then parameter estimates and variances specific to those surveys will
    be combined.
}
