% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_DEPM_Biomass}
\alias{Estimate_DEPM_Biomass}
\title{Combined DEPM parameters to calculate spawning biomass, total number of females and the number of females per weight class using the standard
DEPM approach based off the Parker equation.}
\usage{
Estimate_DEPM_Biomass(adult.pars, adult.vars)
}
\arguments{
\item{adult.pars}{The output of `combine_estimates` with mean Weight and mean F included. This included each of the DEPM parameters estimated using this package.
These can be grouped by Timestep/Region and will be used to estimate results for each combination of these}

\item{adult.vars}{The output of `combine_variances`with var Weight and var F inlcuded. This included each of the DEPM parameters estimated using this package.
These can be grouped by Timestep/Region and will be used to estimate results for each combination of these}
}
\value{
A list of results for spawning biomass and total number of females. These list elements
    will be grouped by Timestep/Region if these were provided with the parameters. The standard deviations of these estimates are also returned
}
\description{
Combined DEPM parameters to calculate spawning biomass, total number of females and the number of females per weight class using the standard
DEPM approach based off the Parker equation.
}
