% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_DEPMWt_biomass}
\alias{Estimate_DEPMWt_biomass}
\title{Combined DEPM parameters to calculate spawning biomass, total number of females and the number of females per weight class using the DEPMwt approach}
\usage{
Estimate_DEPMWt_biomass(adult.pars, adult.vars, weight.pars.vars)
}
\arguments{
\item{adult.pars}{The output of `combine_estimates`. This included each of the DEPM parameters estimated using this package.
These can be grouped by Timestep/Region and will be used to estimate results for each combination of these}

\item{adult.vars}{The output of `combine_variances`. This included each of the DEPM parameters estimated using this package.
These can be grouped by Timestep/Region and will be used to estimate results for each combination of these}

\item{weight.pars.vars}{A dataframe that must include the columns: "Wt_bin", "Prop", "Prop_var", "Wt", "Fecundity", "Fec_var". These correspond
to weight bin number (must start at 1), proportion of females in each weight bin, the multinomial variance of the proportion of females
in each weight bin, weight of each weight bin, fecundity-at-weight of that weight bin and the variance of fecundity-at-weight for that
weight bin, respectively. Wt, fecundity and Fec_var can be calculated using the `Estimate_batch_fecundity` function. These can be
grouped by Timestep/Region and will be used to estimate results for each combination of these}
}
\value{
A list of results for spawning biomass, total number of females and the number of females per weight class. These list elements
    will be grouped by Timestep/Region if these were provided with the parameters. The standard deviations of these estimates are also returned
}
\description{
Combined DEPM parameters to calculate spawning biomass, total number of females and the number of females per weight class using the DEPMwt approach
}
