% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_Mean_Ratio}
\alias{Estimate_Mean_Ratio}
\title{Generic ratio estimator
The internal function used to estimate spawning fraction and sex ratio}
\usage{
Estimate_Mean_Ratio(data, incl.extended.ratio.est = FALSE)
}
\arguments{
\item{data}{A data.frame with numeric columns named "Top" and "Bottom" which represent the numerator and denominator, respectively}

\item{incl.extended.ratio.est}{A logical statement that determines whether the extended ratio method is also returned}
}
\value{
A data.frame that includes the ratio estimate, variance, standard error and coefficient of variation (CV). The
    extended ratio test is also included if requested.
}
\description{
Generic ratio estimator
The internal function used to estimate spawning fraction and sex ratio
}
