% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_mean_W_F}
\alias{Estimate_mean_W_F}
\title{Estimate mean weight (W) and mean (F) for the standard DEPM approach.}
\usage{
Estimate_mean_W_F(
  weight.data,
  TotalWt,
  GonadFrWt = NULL,
  Time = NULL,
  Region = NULL,
  parameters = NULL
)
}
\arguments{
\item{weight.data}{Adult data used to calculate the mean total female weight (W). This data set should ONLY include
females and should include variables for Total weight and/or gonad free weight (Total weight - Gonad weight). These
are used to calculate mean W and provide estimates the intended weight estimates to the fecundity estimator.}

\item{TotalWt}{The column heading in weight.data that represents Total Weight in grams}

\item{GonadFrWt}{The column heading in weight.data that represents Gonad Free Weight in grams. If Gonad weight is not being used
in the fecundity estimator (as total weight is used instead), then this parameter does not need to be specified.}

\item{Time}{A string containing the column name for a timestep if desired as a grouping variable}

\item{Region}{A string containing the column name for a region if desired as a grouping variable}

\item{parameters}{A list of 4 starting parameters for the batch fecundity relationship that
must include: "alpha", "beta", "Sigma0" and "Sigma1"}
}
\value{
A data.frame that includes the Mean W, variance of W, Mean F, variance of F, the covariance of the fecundity-at-weight data and Time and Region if grouping variables were provided
}
\description{
The standard DEPM approach does not use weight bins and does not require a set of weight class parameters.
    However, because of this it requires additional parameters for mean female weight (W) and mean batch fecundity (F).
    These are produced using this function and can be combined with the other weight invariant parameters (P0, A, S and R)
    using the combine_estimates() function. Two data.frames are required which contain weight data and batch fecundity data
    which can be used to estimated each of these parameters.
}
