% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_P0}
\alias{Estimate_P0}
\title{Calculate Daily Egg Production (P0)}
\usage{
Estimate_P0(data, site, Z = NULL, Region = NULL, Time = NULL)
}
\arguments{
\item{data}{A data.frame that contains at least the following columns for Age,  density, and hatching time.
The function can determine these columns based on similarly named variables}

\item{site}{A string containing the column name to idenify the site or sample}

\item{Z}{A numeric vector or single numeric value of egg mortality (Z) value to be used in P0 calculation}

\item{Region}{A string containing the column name for a region if desired as a grouping variable}

\item{Time}{A string containing the column name for a timestep if desired as a grouping variable}
}
\value{
A data.frame of P0 estimates and standard errors for each included Region and Time as well as the values of Z provided
}
\description{
Daily egg production is estimated using mean egg density-at-age for each sample based on a pre-defined
    estimate of egg mortality (Z). If multple years and/or regions are included in the data then a separate estimate of P0 can be
    provided by specifying the columns that correspond to these variables.
}
