% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_proportion_female}
\alias{Estimate_proportion_female}
\title{Estimate_proportion_female}
\usage{
Estimate_proportion_female(data, Weight, max.weight, bin.width, Time = NULL,
  Region = NULL)
}
\arguments{
\item{data}{A dataframe that conatins female weight in grams. No males should be included and must be removed before use}

\item{Weight}{A character string of the variable in `data` that specifies weight in grams}

\item{max.weight}{An integer of the upper weight bin boundary in grams}

\item{bin.width}{An integer for the bin width in grams}

\item{Time}{A string containing the column name for a timestep if desired as a grouping variable}

\item{Region}{A string containing the column name for a region if desired as a grouping variable}
}
\value{
A dataframe with the columns Wt_bin, n, Prop and Prop_var. Timestep and region columns are returned if used
}
\description{
Determine the proportion of of the population in each weight bin. This is performed for females
    only and is an input parameter for the DEPMwt model.
}
