% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Estimate_proportion_female}
\alias{Estimate_proportion_female}
\title{Estimate_proportion_female}
\usage{
Estimate_proportion_female(
  data,
  Weight,
  Weight.bins = NULL,
  Time = NULL,
  Region = NULL
)
}
\arguments{
\item{data}{A dataframe that conatins female weight in grams. No males should be included and must be removed before use}

\item{Weight}{A character string of the variable in `data` that specifies weight in grams}

\item{Weight.bins}{A numeric vector of weights to to create bins. Must be the lower bound for each bin. Midpoints and
upper bounds are automatically determined.}

\item{Time}{A string containing the column name for a timestep if desired as a grouping variable}

\item{Region}{A string containing the column name for a region if desired as a grouping variable}
}
\value{
A dataframe with the columns Wt_bin, n, Prop, Prop_var, Lower_Wt, Upper_Wt, Bin_Width and Mid_point. Timestep and region columns are returned if used
}
\description{
Determine the proportion of of the population in each weight bin. This is performed for females
    only and is an input parameter for the DEPMwt model.
}
